Added Colored Textbox, Port tools, UPnP,
Fixed lots of bugs.
Removed Dev/GCMod protection (as to comply with wiki.vg rules)
Added Option to ignore omnibans for incoming players (Thought it would make all of the issues end)

+++ Commands/CmdClearBlockChanges.cs
@@ -18,39 +18,40 @@
 using System;
 using System.Data;
 using MCForge.SQL;
+using System.Text.RegularExpressions;
 //using MySql.Data.MySqlClient;
 //using MySql.Data.Types;
 
 
-namespace MCForge.Commands
-{
-    public class CmdClearBlockChanges : Command
-    {
-        public override string name { get { return "clearblockchanges"; } }
-        public override string shortcut { get { return "cbc"; } }
-        public override string type { get { return "mod"; } }
-        public override bool museumUsable { get { return false; } }
-        public override LevelPermission defaultRank { get { return LevelPermission.Admin; } }
-        public CmdClearBlockChanges() { }
+namespace MCForge.Commands {
+	public class CmdClearBlockChanges : Command {
+		public override string name { get { return "clearblockchanges"; } }
+		public override string shortcut { get { return "cbc"; } }
+		public override string type { get { return "mod"; } }
+		public override bool museumUsable { get { return false; } }
+		public override LevelPermission defaultRank { get { return LevelPermission.Admin; } }
+		public CmdClearBlockChanges() { }
 
-        public override void Use(Player p, string message)
-        {
-            if (p == null)
-            {
-                Player.SendMessage(p, "This command can only be used in-game");
-                return;
-            }
-            Level l = Level.Find(message);
-            if (l == null && message != "") { Player.SendMessage(p, "Could not find level."); return; }
-            if (l == null) l = p.level;
+		public override void Use(Player p, string message) {
+			if ( p == null ) {
+				Player.SendMessage(p, "This command can only be used in-game");
+				return;
+			}
+			Level l = Level.Find(message);
+			if ( l == null && message != "" ) { Player.SendMessage(p, "Could not find level."); return; }
+			if ( l == null ) l = p.level;
 
-            if (Server.useMySQL) MySQL.executeQuery("TRUNCATE TABLE `Block" + l.name + "`"); else SQLite.executeQuery("DELETE FROM `Block" + l.name + "`");
-            Player.SendMessage(p, "Cleared &cALL" + Server.DefaultColor + " recorded block changes in: &d" + l.name);
-        }
-        public override void Help(Player p)
-        {
-            Player.SendMessage(p, "/clearblockchanges <map> - Clears the block changes stored in /about for <map>.");
-            Player.SendMessage(p, "&cUSE WITH CAUTION");
-        }
-    }
+			if ( !Regex.IsMatch(l.name.ToLower(), @"^[a-z0-9]*?$") ) {
+				Player.SendMessage(p, "Level name is not accepted");
+				return;
+			}
+
+			if ( Server.useMySQL ) MySQL.executeQuery("TRUNCATE TABLE `Block" + l.name + "`"); else SQLite.executeQuery("DELETE FROM `Block" + l.name + "`");
+			Player.SendMessage(p, "Cleared &cALL" + Server.DefaultColor + " recorded block changes in: &d" + l.name);
+		}
+		public override void Help(Player p) {
+			Player.SendMessage(p, "/clearblockchanges <map> - Clears the block changes stored in /about for <map>.");
+			Player.SendMessage(p, "&cUSE WITH CAUTION");
+		}
+	}
 }
